requirements.txt
python-dotenv
langchain-google-genai
streamlit
mcp
langchain-mcp-adapters
langgraph
pypdf
fpdf2
